Deletando classe que nao estava sendo usada e refatorando para melhorar retorno para api caso nao encontre recurso

+++ src/main/java/br/com/guilherme/alura/forumalura/controller/TopicosController.java
@@ -2,6 +2,7 @@ package br.com.guilherme.alura.forumalura.controller;
 
 import java.net.URI;
 import java.util.List;
+import java.util.Optional;
 
 import javax.transaction.Transactional;
 import javax.validation.Valid;
@@ -58,9 +59,12 @@ public class TopicosController {
     }
     
     @GetMapping("/{id}")
-    public DetalhesDoTopicoDTO detalhar(@PathVariable("id") Long id) {
-    	Topico topico = topicoRepository.getOne(id);
-    	return new DetalhesDoTopicoDTO(topico);
+    public ResponseEntity<DetalhesDoTopicoDTO> detalhar(@PathVariable("id") Long id) {
+    	Optional<Topico> topico = topicoRepository.findById(id);
+    	if(topico.isPresent()) {
+    		return ResponseEntity.ok(new DetalhesDoTopicoDTO(topico.get()));
+    	}
+    	return ResponseEntity.notFound().build();
     }
     
     @PutMapping("/{id}")
@@ -68,15 +72,24 @@ public class TopicosController {
     public ResponseEntity<TopicoDTO> atualizar(@PathVariable("id") Long id,
     										   @RequestBody @Valid AtualizacaoTopicoFormDTO form) {
     	
-    	Topico topico = form.atualizar(id, topicoRepository);
-    	return ResponseEntity.ok(new TopicoDTO(topico));
+    	Optional<Topico> optional = topicoRepository.findById(id);
+    	if(optional.isPresent()) {
+        	Topico topico = form.atualizar(id, topicoRepository);
+        	return ResponseEntity.ok(new TopicoDTO(topico));
+    	}
+    	return ResponseEntity.notFound().build();
     }
     
     @DeleteMapping("/{id}")
     @Transactional
     public ResponseEntity<?> remover(@PathVariable("id") Long id) {
-    	topicoRepository.deleteById(id);
-    	return ResponseEntity.ok().build();
+    	
+    	Optional<Topico> optional = topicoRepository.findById(id);
+    	if(optional.isPresent()) {
+    		topicoRepository.deleteById(id);
+        	return ResponseEntity.ok().build();
+    	}
+    	return ResponseEntity.notFound().build();
     }
 }
